fix(storage): timestamps, PG/MySQL column widths, and the claim/release model
Four defects, one deletion. Every claim below was verified against live
PostgreSQL 16 and MySQL 8.4, not inferred.

1. Timestamps were bound as TEXT into TIMESTAMP columns (PostgreSQL, 42804)

   BindVal had no datetime variant, so `updated_at: Some(v)` was stringified and
   bound as &str, which sqlx declares as oid 25. PostgreSQL has no assignment cast
   from text to timestamp. Every update carrying an explicit updated_at failed —
   i.e. every BRC-40 sync merge into a PostgreSQL store, 12 call sites in
   process_sync_chunk.rs. Added BindVal::DateTime and bound natively, following
   the pattern insert.rs already used. The same fix applies to find.rs's `since`
   filter: a bound TEXT parameter is not an untyped literal, so `timestamp >= text`
   cannot resolve either. SQLite still writes its TEXT format; existing files
   depend on it.

2. Column widths that could not hold their own values

   commissions.satoshis was INTEGER (PG) / INT (MySQL) while the model is i64.
   outputs.providedBy was VARCHAR(10) in MySQL after 003 narrowed it based on a
   two-variant ENUM, but the Rust enum has three and the third serialises as
   "you-and-storage" — 15 characters. Under strict mode that is ER_DATA_TOO_LONG;
   otherwise it truncates silently and the next read fails to parse the enum.

3. The claim guard refused legitimate spends

   mark_inputs_spent guards on `spentBy IS NULL`, but the candidate query filters
   only on `spendable`, and two signer paths restored inputs to spendable while
   leaving spentBy pointing at the dead transaction — broadcast_outcome.rs and
   signer/methods/abort_action.rs, where four other release paths cleared it.
   After any failed broadcast or signer abort, those UTXOs became permanently
   unspendable. Reproduced deterministically, single-threaded; mutation-confirmed
   the guard was the cause. Fixed at the root: releases now go through one method
   that clears spentBy, so `spendable = 1 AND spentBy IS NOT NULL` is no longer
   reachable rather than merely tolerated.

   That release carries the guard TS lacks entirely:

       WHERE outputId IN (...) AND (spentBy = ? OR spentBy IS NULL)

   Release what this transaction claimed, and what nobody claimed, but never what
   another transaction claimed. TS's releaseInputsAllocatedToFailedTransaction
   matches on (txid, vout) regardless of spentBy and can release a live claim.
   The IS NULL arm is not laxity — it is the rawTx recovery path, where a
   transaction is authoritative about inputs whose claim was never persisted.

4. The replan loop is deleted

   Neither reference has one. Go returns ErrUTXOContention and lets the whole DB
   transaction abort (repo/transactions.go:115); TS has no such situation. Ours
   was the direct cause of two further defects — partial claims were never
   released before replanning, so a later successful attempt committed UTXOs
   marked spent by a transaction that did not spend them; and a short count was
   blamed on noSendChange by plan membership rather than by which row actually
   failed. A short claim now returns an error and the enclosing transaction rolls
   back, which maintains for free the invariant the compensation code was
   maintaining by hand. 235 lines to 174, and the two-mode release parameter this
   briefly needed is gone with it.

Coverage. Before this branch the server suites executed 0 UPDATE statements —
measured from server-side statement logs, which is why a total PostgreSQL failure
sat undetected behind a green suite. They now execute 19 across 16 tables.

Verified: SQLite 961/0, PostgreSQL 51/0, MySQL 51/0, clippy -D warnings and fmt
clean. No test present on main was removed. Mutation-checked in both directions:
reverting the timestamp bind reproduces 42804, and reverting the signer release
turns the spentBy assertions red.

+++ migrations/postgres/001_initial.sql
@@ -108,7 +108,7 @@ CREATE TABLE commissions (
     "commissionId" BIGSERIAL PRIMARY KEY,
     "userId" BIGINT NOT NULL REFERENCES users("userId"),
     "transactionId" BIGINT NOT NULL UNIQUE REFERENCES transactions("transactionId"),
-    satoshis INTEGER NOT NULL,
+    satoshis BIGINT NOT NULL,
     "keyOffset" VARCHAR(130) NOT NULL,
     "isRedeemed" BOOLEAN NOT NULL DEFAULT FALSE,
     "lockingScript" BYTEA NOT NULL
